Fixed bad indexing of qal array in eparas, line 125
If the internal default enpara file was used it could happen that
index 3 of the qal array in eparas, line 125 was wrong. This can be
tracked back to a missing initialization of enpara%skiplo. I now
initialize this array in inpeig.f90 before default_enpara is called.
The bug fix is in analogy to the initialization of skiplo in
atom_input where an enpara file is generated by inpgen.