Initial introduction of the loading of inp.xml by fleur
For very simple test systems fleur can now use the inp.xml file
as an input. Note, however, that the functionality is still
very incomplete. Therefore, preliminary fleur has to be invoked
with -xmlInput to make use of inp.xml.

Besides the feature incompleteness it is expected that there are
still some bugs in loading inp.xml. So far it is
only tested for Si and Fe. For Fe a small deviation from the
reference total energy is observed (0.03 mHtr), the magnetic
moments are off by 0.003. The cause is yet unknown.