Fixed bug in setup of HDLOs in r_inpXML
r_inpXML included a call to setlomap. This subroutine has the side effect
to reinitialize ulo_der to a value according to the definition in the old
inp file. I removed the call to setlomap and included the missing other
initializations directly in r_inpXML.

Note that there is still a bug in the LO setup iff multiple LOs with same
l quantum numbers are defined for the same species. Fleur has the rquirement
that the LOs are ordered with respect to the l quantum numbers. This ordering
is not yet performed in r_inpXML.