Fixed bug in initialization of vacuum%layerd
Even if not vacuum DOS had to be calculated the corresponding
line in inp.xml was read in. This overwrote the initialization
of vacuum%layerd (in the here observed case to 0). The possible
entries for this variable should be more restricted by the Schema
file.

Also in this commit: Added copyright notice to io/r_inpXML.F90