Added FleurOutputSchema.xsd to io/xml/
Note: The planned usage of XML Schema any elements is not possible
as it makes 'the content model nondeterministic'. Therefore this
schema is stricter than planned. It is also incomplete at the
moment, e.g., orbital magnetic moments and forces are not yet
covered.

Also in this commit some smaller layout changed in out.xml.