Added Bi2Te3XML test. Works.
Note that the XML version of this test produces other
results in comparison to the conventional test. This
is due to a different number of eigenvalues that are
calculated by default and that enter the second variation
for the SOC contributions. For the XML code there is a
minimum number of these eigenvalues that is larger than
the number calculated by default with the conventional
FLEUR. This reason for the differences has been verified.

Note also that the XML version of this test seems to need
considerably less runtime. The reason for this is unknown.