Probably corrected a bug in hlomat call.
I compared the calls to hlomat to those of an old
reference fleur version and realized that depending on the
call some array indices were different. I corrected them such
that they correspond to the old version. None of the test
cases seems to be affected by this so maybe a test case should
be written that is sensitive to these calls.

To me the discrepancies to the old reference fleur version
looked like a copy&paste error when the calls were rewritten
with the new types. Therefore I corrected it.