Fixed array index bug in eigen/tlmplm_store.F90
With strict debugging options I got the error that index 3
of the array tuulo had a value larger than the upper bound. The
array indizes were not explicitely provided at that point and
that surely was wrong. I fixed that. The error is gone. However,
I don't understand how the actual error was caused by this.

Maybe there is still something else.