Fixed erroneous generation of additional identity symmetry matrices
Time reversal symmetry is exploited to increase the number of symmetry
operations used for the generation of the Brillouin zone iff no inversion
symmetry is present. So far the inverse of each symmetry operation was
added, ignoring translation vectors. However, this could lead to further
identity symmetry operations which caused problems as this matrix has
to be excluded for several code parts.

The fix was developed in collaboration with D. Wortmann.