Reduce the number of digits written to inp.xml to 10
For some numbers more than 10 digits were written out to inp.xml.
But it turned out that starting with the 11th digit numerical
noise becomes relevant so that the numbers became strange. It is
surprising that we already have numerical noise in the 11th digit.
I expected this to happen at digit 15 or so.