Introduced alternative construction of IBZ.
Since the routine brzone featured many bugs that occured in special
cases I now reprogrammed the functionality of the routine in the
new subroutine brzone2. The new algorithm is simpler and should
be less vulnerable to the mentioned special cases. It also is slightly
slower. The new routine takes about 2 seconds.

By commenting the new routine out and the old one in both algorithms
can directly be interchanged.