Calculate a save value for nbsz in init/brzone2.f90
nbsz determines the size of the region in reciprocal space considered
to construct the IBZ. Up to now this variable was set to 3 in julia.f.
But this value is not necessarily save. We now calculate a save but
also conservative value for this now local valiable nbszLocal in
brzone2.f90.

Also in this commit: Added build.* to .gitignore.