Bugfix: Made Broyden mixing of the density matrix work again
The Broyden mixing for the density matrix was broken because
in the routine main/mix.F90 one of the two calls to brysh1 was wrong.
In both calls the old density matrix was passed to brysh1, but in
one of the calls this should be the new density matrix. This is
corrected now.