Pushed up a threshold in init/brzone2.f90
The threshold was used at several places to decide whether two values
are identical, on which side of a plane a point lies, and so on. The
strict value before lead to problems as some points were not considered
to be identical due to numerical reasons, although they were meant to
be identical. The larger threshold now solves the problems.

However, if more problems like this arise it should be checked whether
the threshold has to have different values at different places in the
code. Also additional code may be required to treat certain cases in
a special way.