Adapted CuDOSXML test to recent move from the brzone to the brzone2 routine
In brzone2 the IBZ is constructed with a different algorithm in comparison
to brzone. The outcome is that the corners and faces of the IBZ are provided
in a differnt order. Note that there is no clear order for these lists
predefined. The k-point mesh generated by kpttet slighly relies on the order
of the faces and corners. So the usage of brzone2 lead to the generation of
a slighly different but also correct k-point mesh.

The energy mesh in the DOS.1 file depends on the k-point mesh. Thus the
energy grepped for in the CuDOSXML test had to be adapted to the new mesh.