Changed output slightly to correct output conversion errors
At some places input files for the input file generator lead
to "output conversion errors" in the debug version of inpgen, i.e.,
instead of numbers something like *** was written out. This happened
if the input contained too many atoms or the unit cell was too
long in one of the dimensions. I added a few digits to the output
at the respective places to overcome this problem.