Quick fix for wrong times in juDFT_times.
I put the resetIterationDependentTimers call to the beginning of
each iteration. This should provide correct timings in
juDFT_times if a single iteration is performed. Right now, the
iteration dependent times in juDFT_times are not cumulated over
the iterations. Only the total runtime is for all iterations.
The other times are only for the last iteration performed.

If the timings for each iteration are relevant they can be
found in the out.xml file.