Introduced wrapper for core density IO and applied it to some files
Note: There is a change in this commit that may or may not fix
or produce a bug that might be there or not: The rhc array in the
subroutine signature of cored was missing the spin dimension that
it has in cdngen. This is now corrected and the array is now correctly
filled in cored. It is unclear if this changes the behavior of the
code at some place. It might have an effect in cdnovlp as this array
also goes there.