Read in minDistance from inp.xml file
Now the attribute minDistance is read in from the input file into
the variable input%minDistance. By default this variable is 0.0 but
it can be set to finite values in the inp.xml file. It is supposed
to be an additional optional convergence criterion to stop calculations
when they are converged.