Reverted last commit on global/chkmt.f90 and corrected the related bug.
The bug actually appeared because the Bravais matrix was not read in
correctly. The scaling was applied in a wrong way.

The bug affected the size of the MT radii. Therefore I first assumed
that it was related to global/chkmt.f90. It was not. In fact, I wrote
super code in global/chkmt.f90. :)