Introduce the possibility to delete densities from cdn.hdf
The feature is used by invoking fleur with -hdf_cdn -dd a-b
to delete all densities with index between a and b. Using
-dd a only deletes density a.

Note: The required space for the densities will not yet be given
back to the system. The size of the file stays the same. A
solution to this problem may be the usage of the tool h5repack.
It should be available everywhere.