Removed inconsistent usage of scale in inpgen/setab.f
In setab.f There was a problem that for some use cases "scale" had to
be index in a certain way while in others in a different way. I removed
these different cases by directly multiplying the elements of the lattice
vectors a1, a2, a3 in inpgen/lattice2.f with the a, b, c parameters. For
the cases when inpgen/lattice2.f is used the scale vector now is 1.0
after the routine.