Reverted changes to setab.f back to the last changes to this routine from Gustav
I originally made the changes because the scale parameters acted differently
than in version 0.26. It turned out that the behavior in 0.26 is wrong, i.e.,
it does not comply with the documentation.

The scale parameters act on the x, y, and z axes. They do not act on the
lattice vectors.