Move qvac,qvlay allocation back from cdnval to cdngen
Someone moved the allocation of qvac and qvlay from cdngen to cdnval.
This made problems for noco film calculations. In this case it is
important that these arrays are allocated and set to zero outside the
spin loop directly around cdnval.

The problem that actually occured was a wrong writing of the enpara
file. because pvac and svac were 0 in cdnval, mix_enpara.

Please note: There are more arrays that were moved from cdngen to
cdnval. Maybe it should be checked at some point whether there is
another, similar problem due to this.