Corrected theoretically possible segmentation fault in io/xmlIntWrapFort.f90
The issue was due to the reading of a c string into a fortran character array.
Up to now all 255 elements were read into the array and later the output
was reduced to the part in front of the c string termination null character.

Of course, this is wrong and valgrind complained. Everything after the string
termination character might not belong to the program. Therefore I fixed this.
But to my knowledge this issue never produced a problem.