Some initializations that may prevent writing garbage to the garbage section of cdn.hdf
valgrind reports that some uninitialized bytes are in a buffer to be written out to cdn.hdf.
The bytes surely are in that part of the buffer that does not contain relevant data. To me
The reported line for the error source is the close operation for the file. Therefore
it is not clear whether this originates from fleur or from HDF5. I can only identify that
this appeared while writing out the stars. Therefore I initialized every associated array
that did not seem to be initialized before setting the relevant elements to certain values.

I had to start valgrind with --num-callers=25 to see at least the fleur context for the
reported missing initializations.