Comment out "test_fileexists" from Fe_bct_LO test
Somehow the testcase was ended due to the now commented line if the files
were not found. As it was ended, the succeeding validations could not be run
and the test passed. But it should have failed.

This was relevant for the MPI eigenvector parallelization for this test case.
This is not implemented and therefore has to fail.