Revert changes in NiO_ldauXML test and bugfix in io/r_inpXML.F90
The fixed bug was the ignoring of the ldaU XML element if the version number
of the inp.xml file was not 0.29. For the NiO_ldauXML test this meant that
no linear mixing was used for the density matrix.

Please note that there are still slight differences between the results
obtained with this version and the last release version. I changed the
reference value for the total energy from -3191.9717 to -3191.9707.