Quick fix for Fe_bct test
The test failed because the unrotated input density was not kept for the mixer.
I solved it by introducing an inDenRot variable that is a copy of inDen and then
gets rotated in rhodirgen if that soubroutine is in the program flow. It is only
used as an input to those subroutines that require the rotated input density or
are insensitive to it.