Quick fix for MPI trouble in MPI:Fe_bct test
The problem was that the 2nd dimension of vCoul%pw was set to 2 by the assignment
modified in this commit. This lead to an MPI broadcasting problem as it was only
done for rank 0. At some point one should check whether we actually need more
than the 2 spin dimensions for this potential contribution.