Fix OpenMP troble in eigen/hsmt_ab.f90
The problem was a "double free or corruption" error. It was related to the
ylm array in eigen/hsmt_ab.f90 being allocated on the heap and not
being a stack array. It seems to be a requirement that OpenMP private
variables are stack variables. At least putting ylm on the stack
seems to have fixed the problem.