Some cleanups in cdn/cdnval.F90 and related files
Note: This cleanup includes the inclusion of the qvlay array
into the t_regionCharges type. This breaks the program logic
for the noco%l_noco case as it is created inside a certain
spin loop but should be created outside this loop. It is not
yet clear how this will be solved. Probably the regCharges
type should just move to cdngen and the arrays in it should
have full spin dimensions.