Fixed possible bug in setting the XC functional in r_inpXML
The problem was that the output variable namex wasn't set any more. Of course,
this variable is not used for most things but it is at least used in w_inpXML
which is called at several points in the code. At some point this variable
should be eliminated in the whole code.