Introduce Logical Variable yukawa_residual
At the moment this variable is without function. Later it will be used as a switch between the Yukawa and Coulomb cases in the potential generation.
Since the preconditioning parameter is now stored in the input file and since both the Coulomb and Yukawa potentials are generated in a calculation with preconditioner, the switch cannot simply be based on the value of the preconditioning parameter. This is quite ugly and we might want to change this somehow later.