Commented out OpenMP parallelization in vgen/mpmom.F90 (again)
I already commented this out in commit 415885af7af8a8cb024625a23f19e3373adb3628.
I don't know who commented it in again, but the problem with this parallelization is
not solved. With ifort 18 I obtain for the CuBulkXML test with this parallelization
a qlmp(0,0,1) of (1.490549732270697E+307,-2.695127599127286E-033). Without the
parallelization it looks okay: (0.990659014383078,-2.695127599127286E-033). It may
be that only the 0,0,1 element is affected. But this should be checked before the
parallelization is commented in again.