Quick fix for issue #178.
I fixed the issue by reducing a threshold in inpgen/bravais_symm.f. This
leads to less symmetry operations that are found if the specified Bravais
matrix is not precise enough. I hope this does not prevent the detection
of symmetry operations that we clearly want to find. Otherwise we have
to find a more sophisticated fix for the issue.