Fix vx in vgen/vgen_xcpot.F90
Outside of vgen/vgen_xcpot.F90 this potential contribution is used such
that it is assumed that every entry of the interstitial part is divided
by the number of G vectors contributing to the respective stars.

This fixes an issue in the hybrid functionals calculation. Also for
GW it is needed in this form.