Cured segmentation fault error symptom in cdnval
In some calculations a segmentation fault occured when the we and eig arrays in
cdnval were allocated. No reason for this was identified. Every parameter was
harmless. To avoid the occurence of this error at this place I just moved
these arrays to the stack. It seems to help and up to now no other error at a
different place occured.