Make Fleur compilable again (on MY machine)
This commit should not change anything, but it does.
The Intel compiler seems to be sensitive on the order and position
of use statements. Without this change the compilation of Fleur
hangs at several places. I don't have a general principle on
how and where the use statements should be, but:

1. more general use statements should be before use statements that
themselves use the more general one.
2. use statements should be simple. No ", only ..."
3. Less use statements is better. If there are several subroutines
in a module put the use statements generally into the module and not
into the subroutines.

The reality, however, seems to be slightly more complex.

This might also be connected to the underlying hardware resources.