Try to translate memory problem in MPI_PUT call in io/eig66_mpi.F90 into an error
For the investigated case this change will move the error from
the MPI_PUT call to a segmentation fault in the related allocate
statement. This is not the wanted behavior as there is still no
Fleur error thrown. But this behavior might depend on the Fortran
compiler, the size of the related arrays and so on. So this commit
is done with the hope that it leads to the wanted behavior for some
cases.