Add first version of automated parser tests (Only works at the moment if masci tools develop branch is cloned manually)
The names of tests and their respective files using the execute_fleur fixture are collected in a set during discovery and used to parametrize the parser tests.
This assumes that the parser tests are discovered after the fleur tests, which is True at the moment since it goes through the folders alphabetically

Kind of a dirty hack but this is the only way I found to use the test_names themselves as parameters for the tests
The small pytest-dependency plugin is used to make sure that parser tests are only run if the corresponding fleur test passed
For this to work each fleur test has to be marked with @pytest.mark.dependency(). This is automatized via the config option automark_dependency