Remove sign in optional/plot.F90 to fix different magnetization directions in MT, interstitial
Note: This sign seems to come and go from version to version. It is not
clear whether this choice is correct. It is at least not consistent with
the other regions. But a consistency between the regions would have to
be fixed in other parts of the code.

To me it is not clear whether there is a real problem related to this.