Allow modification of the mpi command in the python test system while keeping the freedom for tests of enforcing a set number of mpi processes. (See #637)
The juDFT_MPI variable should now no longer specify the number of processes. Either the argument for number of processes is put last and the number will be appended (e.g `mpirun -n `) or the text {mpi_procs} is inserted instead of the number of processes (e.g. `mpirun -n {mpi_procs} --allow-run-as-root`). The number of processes can be passed to the `execute_fleur` fixture with the argument `mpi_procs`. A number of mpi processes for the whole test session can be enforced by setting the `juDFT_MPROCS` variable