Some more fixes to spin rotations of density matrix.
This is only used with the lflip feature for now since the general usage with
the l_relaxSQA switch leads to non-convergence (probably because of some inconsistency in the mixing or
rotation). This means that the density matrix lags a bit behind the rest of the density in the relaxation
of the spin frame. It will act as if the same density matrix is used in both frames. One reason for the
worse convergence could be that the offdiagonal parts get smaller when the density matrix is adjusted,
meaning the force pulling atoms to the relaxed SQA is smaller. But this has to be investigated further