Hide traceback in execute_inpgen and execute_fleur fixtures
This way when Fleur encounters an error the test output is much more clear
for people not familiar with the internals of the test system

One drawback of this is if there is a genuine exception with e.g. copying
the files to the work directory the traceback will also be hidden

Since this should not happen and should in any case lead to a nice
understandable error message the change is worth it. If the tracebacks should
be shown a new flag ``--show-tracebacks`` is introduced to show all
of the tracebacks