Reduce amount of yml specification needed to use the python test system
Introduced template .test-fleur and .test-fleur-not-on-push for tests on every commit and
in the nightly pipeline respectively. Allow configuration of using only certain markers or
changing configuration.

Usage is done by using the extends keyword in the concrete tests and overwriting the needed variables

The run_tests.sh script also does the tee command previously done in the CI

Also put in an additional test before the whole test suite is run only running tests marked with fast
to quickly exit if something is fundamentally wrong