Use serial hdf5 build for AiiDA tests
This build is only done on the nightly builds. The reason for this is that the
pytest plugin for running the workflow regression tests has no support for MPI
executables yet and even if they are called without MPI they can lead to strange test
failures. We don't use the non-hdf5 build since the hdf5 build leads to relevant
behaviour changes in aiida-fleur and are assumed to be the main way of using aiida-fleur