Adjust LDA+U tests
Unfortunately the kpoint set has to be a bit larger to see something
testable for the magnetic moment/bandgap for NiO. The test previously
convereged into a completely non-magnetic state making it impossible to
test if the different double countings work as expected. The number of
iterations is limited to keep the runtime down. Added a NiO test with
FLL double counting

The GaAsMultiUForceXML test is adjusted since the last commit switches
of the initial density matrix by default
For the Greensfunction the initial guess is switched on since for
rare-earths the initial density matrix works very well