Some performance improvements to the JEDI spack recipe.
In my test benchmark this gives a speedup of 8%. Remaining issues are: 1. It is more problematic and not yet solved to adapt the cuda architecture for ELPA to "90". 2. Parts of ELPA seem to get compiled with the gnu C++ compiler "cpp". Therefore the "-tp=host" had to be removed for the cppflags. We have to find out how to compile everything with NVHPC.